requirement1.txt
fastapi
uvicorn[standard]
torch
torchvision
pillow
numpy
opencv-python-headless
